requirements.txt
fastapi
uvicorn
requests
mistralai
cohere
openai
openai
dotenv
pydantic
python-multipart
PyPDF2
google-cloud-translate